requirements.txt
numpy
pandas
streamlit
mysql-connector-python
scikit-learn
joblib
imbalanced-learn
passlib